+++ docs/CHANGE_LOG.md
@@ -5,3 +5,4 @@
 - Improved the ignore check to use actual boolean value.
 - Added improved checks around the elements (incase they are removed from the DOM).
 - Added a new configuration option called "assignMissingIds" (defaults to false), which will assign Ids to the elments that do not contain them.
+- Added a new configuration option called "elementIdPrefix" (defaults to "rattribute"), which states the element Id prefix to used when "assignMissingIds" is enabled.